feat: rules_oci support
This commit adds oci_images as a new element on k8s resources. Labels
should point to oci_image rules. The resolver will read the image digest
from the index.json file of each oci_image and replace the image
references in the yaml.

Limitations: currently only either oci images or docker images are
supported. This is just an artificial limitation and could probably be
removed.